image/requirements.txt
requests
numpy
Pillow
opencv-python
psycopg2-binary
torch
torchvision
transformers
ultralytics
python-dotenv
